refactor: [OWL-00009] Updated database models
