+++ src/enums/games.ts
@@ -1,3 +1,4 @@
 export enum GameCode {
+  Blackjack = 'blk',
   Wordle = 'wdl',
 }
