+++ src/interfaces/games/blackjack.ts
@@ -1,0 +1,5 @@
+export interface BlackjackStats {
+  totalBlackjack: number;
+  totalPlayed: number;
+  totalWon: number;
+}
