+++ src/interfaces/games/index.ts
@@ -1,0 +1,2 @@
+export * from './blackjack';
+export * from './wordle';
