+++ src/interfaces/games/wordle.ts
@@ -1,0 +1,7 @@
+export interface WordleStats {
+  currentStreak: number;
+  distribution: number[];
+  maxStreak: number;
+  totalPlayed: number;
+  totalWon: number;
+}
