--- src/services/statistics.ts
@@ -1,19 +1,0 @@
-import { log } from '@/discord/helpers';
-import { LogCode } from '@/enums/logs';
-import { StatsDocument } from '@/interfaces/statistics';
-import { StatsModel } from '@/models/statistics';
-
-export const deleteStats = async (
-  id: string
-): Promise<StatsDocument | null> => {
-  try {
-    const deleted = await StatsModel.findOneAndDelete({ discord_id: id });
-    return deleted;
-  } catch (error) {
-    log({
-      type: LogCode.Error,
-      description: JSON.stringify(error),
-    });
-    return null;
-  }
-};
